Stand Ende der Vorlesung

+++ Simulator PIC16F84/ProgramMemoryByte.cs
@@ -8,14 +8,14 @@ namespace Simulator_PIC16F84
 {
     public class ProgramMemoryByte
     {
-        private byte value;
+        private int value;
 
         public ProgramMemoryByte()
         {
             value = 0;
         }
 
-        public byte Value {
+        public int Value {
             get { return value; }
             set { this.value = value; }
         }
